Добавлена возможность генерации расписания. Обновлен экран с расписанием.

+++ app/src/main/java/com/example/schreduler/utils/JsonMapConverter.kt
@@ -6,13 +6,13 @@ import com.google.gson.reflect.TypeToken
 
 class JsonMapConverter {
     @TypeConverter
-    fun fromString(value: String): Map<Int, List<Int>> {
+    fun fromString(value: String): Map<Int, List<Long>> {
         val mapType = object : TypeToken<Map<Int, List<Int>>>() {}.type
         return Gson().fromJson(value, mapType)
     }
 
     @TypeConverter
-    fun fromMap(map: Map<Int, List<Int>>): String {
+    fun fromMap(map: Map<Int, List<Long>>): String {
         return Gson().toJson(map)
     }
 }
